Encryption and Decryption
This is program that encrypts and decrypts string. It uses the inputted
string and converts all the characters into ASCII reference values. The
user will then be prompted to enter in a 3 digit number which will
shift the ASCII reference values accordingly. For example, the number
152 for the string “abc” will be encrypted into “bge” because the first
character will be shifted by 1, the second character will be shifted by
5, and the third character will be shifted by 2. In the event that the
shifted character contains ASCII code that does not correspond to a
letter of the alphabet, the ASCII code will be subtracted by 26 to make
sure that it is a letter. When decrypting, the reverse will be done.
The decrypted string will be created when the characters will be
subtracted by the 3 digit number as shown above. In the event that the
ASCII code no longer corresponds to a letter, then the number 26 will
be added to fix the decryption.